Fixed the insert completely and added a new case statement
The insert function should completely work now, I have tested this function many times. I also added a new case, case 5. This case will read in information from a text file, people.txt. The case asked for user input for the name of the file, reads it in, inserts it and then displays the current queue. people.txt is being uploaded as well